+++ ft_strrchr.c
@@ -6,16 +6,16 @@
 /*   By: agaley <agaley@student.42lyon.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/26 13:08:49 by agaley            #+#    #+#             */
-/*   Updated: 2022/11/09 23:48:39 by agaley           ###   ########lyon.fr   */
+/*   Updated: 2022/12/04 23:19:48 by agaley           ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-char	*ft_strrchr(const char *s, int c)
+char *ft_strrchr(const char *s, int c)
 {
-	int		i;
-	char	*s2;
+	size_t i;
+	char *s2;
 
 	s2 = (char *)s;
 	i = ft_strlen(s2) - 1;
